Change all decimals into doubles for larger ranges. The precision is not as equaly important

+++ src/RavenNest.Blazor/Controllers/PlayersController.cs
@@ -246,7 +246,7 @@ namespace RavenNest.Controllers
         }
 
         [HttpPost("{userId}/statistics")]
-        public bool UpdateStatistics(string userId, Many<decimal> statistics)
+        public bool UpdateStatistics(string userId, Many<double> statistics)
         {
             return playerManager.UpdateStatistics(AssertGetSessionToken(), userId, statistics.Values);
         }
