+++ src/RavenNest.BusinessLogic/Net/Models/CharacterSkillUpdate.cs
@@ -6,7 +6,7 @@ namespace RavenNest.BusinessLogic.Net
     {
         public Guid CharacterId { get; set; }
         public string UserId { get; set; }
-        public decimal[] Experience { get; set; }
+        public double[] Experience { get; set; }
         public int[] Level { get; set; }
     }
