+++ src/RavenNest.Models/PlayerExpUpdate.cs
@@ -4,6 +4,6 @@
     {
         public string UserId { get; set; }
         public string Skill { get; set; }
-        public decimal Experience { get; set; }
+        public double Experience { get; set; }
     }
 }
