+++ src/RavenNest.SDK/Endpoints/CharacterStateUpdate.cs
@@ -9,7 +9,7 @@ namespace RavenNest.SDK.Endpoints
     {
         public Guid CharacterId { get; set; }
         public string UserId { get; set; }
-        public decimal[] Experience { get; set; }
+        public double[] Experience { get; set; }
         public int[] Level { get; set; }
     }
